chore(hog-charts): revert AlertOverlay and y-formatters moves
These two files import product code (kea, insightAlertsLogic,
IndexedTrendResult, TrendsFilter), which violates the no-kea/no-product-imports
rule for lib/hog-charts. Move them back to scenes/trends/viz/trends-line-chart/
and switch TrendsLineChart's imports back to local paths. Drop the matching
re-exports from lib/hog-charts/index.ts.

The other relocations in this PR (AnnotationsLayer, AnomalyPointsLayer,
derived-series, goalLinesAdapter, dates) stay put — they're already
product-free.

A generic y-axis format extraction is deferred to Stage 3 where it can be
wired into config.yAxis.format.
