+++ frontend/src/scenes/trends/viz/trends-line-chart/trendsAxisFormat.test.ts
@@ -1,6 +1,6 @@
 import { CurrencyCode, TrendsFilter } from '~/queries/schema/schema-general'
 
-import { buildTrendsYTickFormatter } from './y-formatters'
+import { buildTrendsYTickFormatter } from './trendsAxisFormat'
 
 const NBSP = ' '
