+++ frontend/src/scenes/trends/viz/trends-line-chart/trendsAxisFormat.ts
